+++ Tests/SentryTests/SentryTests-Bridging-Header.h
@@ -79,6 +79,8 @@
 #import "SentryTestIntegration.h"
 #import "SentryThreadInspector.h"
 #import "SentryTracer.h"
+#import "SentryTransaction.h"
 #import "SentryTransport.h"
 #import "SentryTransportFactory.h"
+#import "SentryUIViewControllerSanitizer.h"
 #import "SentryUserFeedback.h"